+++ loadtest/.gitignore
@@ -1,0 +1,2 @@
+# Ignore random FF jsons generated from ff_gen.go
+random.json
