--- loadtest/example_flags.json
@@ -1,36 +1,0 @@
-{
-  "flags": {
-    "myBoolFlag": {
-      "state": "ENABLED",
-      "variants": {
-        "on": true,
-        "off": false
-      },
-      "defaultVariant": "on"
-    },
-    "myStringFlag": {
-      "state": "ENABLED",
-      "variants": {
-        "key1": "val1",
-        "key2": "val2"
-      },
-      "defaultVariant": "key1"
-    },
-    "myFloatFlag": {
-      "state": "ENABLED",
-      "variants": {
-        "one": 1.23,
-        "two": 2.34
-      },
-      "defaultVariant": "one"
-    },
-    "myIntFlag": {
-      "state": "ENABLED",
-      "variants": {
-        "one": 1,
-        "two": 2
-      },
-      "defaultVariant": "one"
-    }
-  }
-}
